refactor/5-move-export_calendar-to-hr_addon_settings.py
