+++ hr_addon/hooks.py
@@ -194,8 +194,8 @@ required_apps = ["hrms"]
 
 doc_events = {
     "Leave Application": {
-        "on_change": "hr_addon.hr_addon.api.export_calendar.export_calendar",
-		"on_cancel": "hr_addon.hr_addon.api.export_calendar.export_calendar"
+        "on_change": "hr_addon.hr_addon.doctype.hr_addon_settings.hr_addon_settings.export_calendar",
+		"on_cancel": "hr_addon.hr_addon.doctype.hr_addon_settings.hr_addon_settings.export_calendar"
     }
 }
