+++ hr_addon/hr_addon/doctype/hr_addon_settings/hr_addon_settings.js
@@ -33,7 +33,7 @@ frappe.ui.form.on('HR Addon Settings', {
 
 	generate_workdays_for_past_7_days_now: function(frm){
 		frappe.call({
-			method: "hr_addon.hr_addon.doctype.hr_addon_settings.hr_addon_settings.generate_workdays_for_past_7_days_now",
+			method: "hr_addon.hr_addon.doctype.workday.workday.generate_workdays_for_past_7_days_now",
 		}).then(r => {
 			frappe.msgprint("The workdays have been generated.")
 		})
